@@ -12,6 +12,7 @@
 namespace spec\Memio\TwigTemplateEngine\TwigExtension\Line;
 
 use Memio\Model\File;
+use Memio\TwigTemplateEngine\TwigExtension\Line\LineStrategy;
 use PhpSpec\ObjectBehavior;
 
 class FileLineStrategySpec extends ObjectBehavior
@@ -20,7 +21,7 @@ class FileLineStrategySpec extends ObjectBehavior
 
     function it_is_a_line_strategy()
     {
-        $this->shouldImplement('Memio\TwigTemplateEngine\TwigExtension\Line\LineStrategy');
+        $this->shouldImplement(LineStrategy::class);
     }
 
     function it_supports_files(File $file)
@@ -28,9 +29,10 @@ class FileLineStrategySpec extends ObjectBehavior
         $this->supports($file)->shouldBe(true);
     }
 
-    function it_needs_line_after_fully_qualified_names_if_file_has_fully_qualified_names(File $file)
-    {
-        $file->allFullyQualifiedNames()->willReturn(array(1));
+    function it_needs_line_after_fully_qualified_names_if_file_has_fully_qualified_names(
+        File $file
+    ) {
+        $file->allFullyQualifiedNames()->willReturn([1]);
 
         $this->needsLineAfter($file, self::IMPORT_BLOCK)->shouldBe(true);
     }
